Reorg tests.
