--- .changes/next-release/feature-Configuration-22255.json
@@ -1,5 +1,0 @@
-{
-  "type": "feature",
-  "category": "Configuration",
-  "description": "A new configuration option cli_timestamp_format was added."
-}
