+++ .changes/next-release/feature-Configuration-61155.json
@@ -1,5 +1,5 @@
 {
-  "category": "Configuration", 
-  "type": "feature", 
-  "description": "Adds a new option to the configuration file to change the timestamp output format displayed by the CLI."
+  "category": "Configuration",
+  "type": "feature",
+  "description": "Adds a new option to the configuration file 'cli_timestamp_format' to change the timestamp output format displayed by the CLI."
 }
